% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{annotCols}
\alias{annotCols}
\title{Get annotation columns}
\usage{
annotCols(mdata)
}
\arguments{
\item{mdata}{matrixData}
}
\description{
Get annotation columns
}
\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotRows<-}}, \code{\link{annotRows}},
  \code{\link{description<-}}, \code{\link{description}},
  \code{\link{main<-}}, \code{\link{main}},
  \code{\link{matrixData-class}}, \code{\link{matrixData}},
  \code{\link{names,matrixData-method}}
}
