% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{annotRows<-}
\alias{annotRows<-}
\title{Set annotation rows}
\usage{
annotRows(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set annotation rows
}
