% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataMethods.R
\name{as.ExpressionSet.matrixData}
\alias{as.ExpressionSet.matrixData}
\alias{as.ExpressionSet.matrixData_}
\title{Converts a MatrixData ExpressionSet}
\usage{
as.ExpressionSet.matrixData(mdata)
}
\arguments{
\item{mdata}{a \code{\link[PerseusR]{matrixData}} object}
}
\description{
Converts a MatrixData ExpressionSet
}
