% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_annotRows}
\alias{create_annotRows}
\title{Create annotation rows}
\usage{
create_annotRows(commentRows, isMain)
}
\description{
Create the annotation rows data.frame from the list
of comment rows parsed from the input file and the main columns indicator
}
