% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{description<-}
\alias{description<-}
\title{Set column description}
\usage{
description(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set column description
}
