% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{imputeData<-}
\alias{imputeData<-}
\title{Set the imputed cells of main data frame}
\usage{
imputeData(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set the imputed cells of main data frame
}
\seealso{
Other matrixData basic functions:
  \code{\link{annotCols}}, \code{\link{annotCols<-}},
  \code{\link{annotRows<-}}, \code{\link{annotRows}},
  \code{\link{description<-}}, \code{\link{description}},
  \code{\link{main<-}}, \code{\link{main}},
  \code{\link{matrixData-class}}, \code{\link{matrixData}},
  \code{\link{imputeData}},
  \code{\link{qualityData}},\code{\link{qualityData<-}},
  \code{\link{names,matrixData-method}}
}
