% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{infer_perseus_annotation_types}
\alias{infer_perseus_annotation_types}
\title{Infer Perseus type annotation row from DataFrame column classes}
\usage{
infer_perseus_annotation_types(df, typeMap)
}
\arguments{
\item{df}{The data.frame}

\item{typeMap}{A list with elements 'Perseus' and 'R'. The ordering determines the mapping}
}
\value{
A vector with perseus type annotations
}
\description{
Infer Perseus type annotation row from DataFrame column classes
}
\seealso{
Based on \code{\link{mapvalues}}
}
