% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\docType{methods}
\name{initialize,matrixData-method}
\alias{initialize,matrixData-method}
\title{matrixData initializer}
\usage{
\S4method{initialize}{matrixData}(.Object, ...)
}
\arguments{
\item{.Object}{Initialized object}

\item{...}{Additional arguments}
}
\description{
Initializes the annotCols data frame to have the
same number of rows as the main data. This might not be the
cleanest solution.
}
