% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{intParamValue}
\alias{intParamValue}
\title{Int parameter value}
\usage{
intParamValue(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The selected number
}
\description{
Extract the value chosen in an \code{IntParam}
}
\examples{
write('<IntParam Name="test_int">\\n<Value>2</Value>\\n</IntParam>', file='tmp.xml')
parameters <- parseParameters("tmp.xml")
intParamValue(parameters, "test_int")
}
