% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{main}
\alias{main}
\title{Get main columns}
\usage{
main(mdata)
}
\arguments{
\item{mdata}{matrixData}
}
\description{
Gets the main collumns (main matrix) of a \code{\link[PerseusR]{matrixData}}
object as a data.frame object
}
\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotCols}}, \code{\link{annotRows<-}},
  \code{\link{annotRows}}, \code{\link{description<-}},
  \code{\link{description}}, \code{\link{main<-}},
  \code{\link{matrixData-class}}, \code{\link{matrixData}},
  \code{\link{names,matrixData-method}}
}
