% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\docType{class}
\name{matrixData-class}
\alias{matrixData-class}
\title{MatrixData}
\description{
MatrixData
}
\section{Slots}{

\describe{
\item{\code{main}}{Main expression \code{data.frame}.}

\item{\code{annotCols}}{Annotation Columns \code{data.frame}.}

\item{\code{annotRows}}{Annotation Rows \code{data.frame}.}

\item{\code{description}}{Column descriptions.}

\item{\code{imputeData}}{Imputation \code{data.frame}.}

\item{\code{qualityData}}{Quality values \code{data.frame}.}
}}

\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotCols}}, \code{\link{annotRows<-}},
  \code{\link{annotRows}}, \code{\link{description<-}},
  \code{\link{description}}, \code{\link{imputeData<-}},
  \code{\link{imputeData}}, \code{\link{main<-}},
  \code{\link{main}}, \code{\link{matrixData}},
  \code{\link{names,matrixData-method}},
  \code{\link{qualityData<-}}, \code{\link{qualityData}}
}
\concept{matrixData basic functions}
