% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\docType{methods}
\name{names,matrixData-method}
\alias{names,matrixData-method}
\title{Get names}
\usage{
\S4method{names}{matrixData}(x)
}
\arguments{
\item{x}{matrixData}
}
\description{
Get the column names of main and annotation columns.
}
