% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{multiChoiceParamValue}
\alias{multiChoiceParamValue}
\title{Multiple choice value}
\usage{
multiChoiceParamValue(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The string representing the value
}
\description{
Extract the value selected in a \code{MultiChoiceParam}.
}
\examples{
write('<MultiChoiceParam Name="test_multi">\n<Value>\n<Item>1</Item>\n<Item>2</Item>\n</Value>\n
<Values>\n<Item>A</Item>\n<Item>B</Item>\n</Values>\n</MultiChoiceParam>', file='tmp.xml')
parameters <- parseParameters("tmp.xml")
multiChoiceParamValue(parameters, "test_multi")
}
