% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{names.matrixData}
\alias{names.matrixData}
\title{Column names of main and annotation columns}
\usage{
\method{names}{matrixData}(x)
}
\arguments{
\item{x}{matrixData}
}
\description{
Column names of main and annotation columns
}
