% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parseParameters}
\alias{parseParameters}
\title{Parse parameters}
\usage{
parseParameters(paramFile)
}
\arguments{
\item{paramFile}{Parameters xml file}
}
\description{
Parse parameters from the parameters xml file.
}
\examples{
write('<IntParam Name="test_int">\\n<Value>2</Value>\\n</IntParam>', file='tmp.xml')
parameters <- parseParameters("tmp.xml")
}
