% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{qualityData<-}
\alias{qualityData<-}
\title{Set the quality values of main data frame}
\usage{
qualityData(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set the quality values of main data frame
}
\seealso{
Other matrixData basic functions:
  \code{\link{annotCols}}, \code{\link{annotCols<-}},
  \code{\link{annotRows<-}}, \code{\link{annotRows}},
  \code{\link{description<-}}, \code{\link{description}},
  \code{\link{main<-}}, \code{\link{main}},
  \code{\link{matrixData-class}}, \code{\link{matrixData}},
  \code{\link{imputeData}}, \code{\link{imputeData<-}},
  \code{\link{qualityData}},
  \code{\link{names,matrixData-method}}
}
