% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{singleChoiceParamInd}
\alias{singleChoiceParamInd}
\title{Single choice index}
\usage{
singleChoiceParamInd(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The index of the value
}
\description{
Extract the index selected in a \code{SingleChoiceParam}.
}
