% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{stringParamValue}
\alias{stringParamValue}
\title{String parameter value}
\usage{
stringParamValue(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The selected string
}
\description{
Extract the value chosen in an \code{StringParam}
}
\examples{
write('<StringParam Name="test_string">\\n<Value>"test"</Value>\\n</StringParam>', file='tmp.xml')
parameters <- parseParameters("tmp.xml")
stringParamValue(parameters, "test_string")
}
