% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.perseus}
\alias{write.perseus}
\title{Write annotated DataFrame in Perseus matrix format}
\usage{
write.perseus(df, outFile)
}
\arguments{
\item{df}{The DataFrame}

\item{outFile}{Path to output file}
}
\description{
Write a DataFrame to file in the custom Perseus matrix file format.
The DataFrame should have a 'annotationRows' attribute set.
}
\seealso{
\code{\link{read.perseus}}
}

