% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{annotCols<-}
\alias{annotCols<-}
\title{Set annotation columns}
\usage{
annotCols(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set annotation columns
}
