% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{annotRows<-}
\alias{annotRows<-}
\title{Set annotation rows}
\usage{
annotRows(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set annotation rows
}
\examples{
mdata <- matrixData(main=data.frame(a=1:3, b=6:8),
annotCols=data.frame(c=c('a','b','c')),
annotRows=data.frame(x=factor(c('1','1'))))
value <- data.frame(y=factor(c('2','2')))
annotRows(mdata) <- value
}
\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotCols}}, \code{\link{annotRows}},
  \code{\link{description<-}}, \code{\link{description}},
  \code{\link{imputeData<-}}, \code{\link{imputeData}},
  \code{\link{main<-}}, \code{\link{main}},
  \code{\link{matrixData-class}}, \code{\link{matrixData}},
  \code{\link{names,matrixData-method}},
  \code{\link{qualityData<-}}, \code{\link{qualityData}}
}
\concept{matrixData basic functions}
