% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataMethods.R
\name{as.ExpressionSet.matrixData_}
\alias{as.ExpressionSet.matrixData_}
\alias{as.ExpressionSet.matrixData}
\title{Converts a MatrixData ExpressionSet

Converts a MatrixData ExpressionSet

function fo convert a \code{\link[matrixData]{PerseusR}} \code{\link[ExpressionSet]{Biobase}}}
\usage{
as.ExpressionSet.matrixData_(mdata)
}
\arguments{
\item{mdata}{a \code{\link[matrixData]{PerseusR}} object}
}
\value{
an ExpressionSet object
}
\examples{
mD <- matrixData(
main=data.frame(a=1:3, b=6:8),
annotCols=data.frame(b=c('a','b','c')),
annotRows=data.frame(x=factor(c('1','1'))))

eSet <- as(mD, "ExpressionSet")
print(eSet)
}
