% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataMethods.R
\name{as.matrixData.ExpressionSet}
\alias{as.matrixData.ExpressionSet}
\alias{as.ExpressionSet.matrixData}
\alias{as.ExpressionSet.matrixData_}
\title{Converts a MatrixData ExpressionSet}
\usage{
as.matrixData.ExpressionSet(x)
}
