% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataMethods.R
\name{as.matrixData.ExpressionSet}
\alias{as.matrixData.ExpressionSet}
\title{Coerces an ExpressionSet into a MatrixData}
\usage{
as.matrixData.ExpressionSet(ExpressionSet)
}
\arguments{
\item{ExpressionSet}{an \code{\link[Biobase]{ExpressionSet}} object}
}
\value{
returns a \code{\link[PerseusR]{matrixData}} object
}
\description{
function to convert an \code{\link[Biobase]{ExpressionSet}} object into a
\code{\link[PerseusR]{matrixData}}
}
\examples{

\donttest{
eSet <- eSet <- Biobase::ExpressionSet(matrix(1:10, ncol = 2))
mD <- as(eSet, "matrixData")
print(mD)
}
}
