% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataMethods.R
\name{as.matrixData.ExpressionSet_}
\alias{as.matrixData.ExpressionSet_}
\alias{as.matrixData.ExpressionSet}
\title{Converts a ExpressionSet MatrixData}
\usage{
as.matrixData.ExpressionSet_(ExpressionSet)
}
\arguments{
\item{ExpressionSet}{an \code{\link[ExpressionSet]{Biobase}}}
}
\value{
a \code{\link[matrixData]{PerseusR}} object
}
\description{
function fo convert an \code{\link[ExpressionSet]{Biobase}} object into a
\code{\link[matrixData]{PerseusR}}
}
\examples{

eSet <- eSet <- Biobase::ExpressionSet(matrix(1:10, ncol = 2))
mD <- as(eSet, "matrixData")
print(mD)
}
