% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{boolParamValue}
\alias{boolParamValue}
\title{Bool parameter value}
\usage{
boolParamValue(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The selected number
}
\description{
Extract the value chosen in an \code{BoolParam}
}
