% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_annotRows}
\alias{create_annotRows}
\title{Create annotation rows}
\usage{
create_annotRows(commentRows, isMain)
}
\arguments{
\item{commentRows}{list of comment rows}

\item{isMain}{logical array indicating all main columns}
}
\description{
Create the annotation rows data.frame from the list
of comment rows parsed from the input file and the main columns indicator
}
\seealso{
used by \code{\link{read.perseus}}
}
