% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{description<-}
\alias{description<-}
\title{Set column description}
\usage{
description(mdata) <- value
}
\arguments{
\item{mdata}{matrixData}

\item{value}{value}
}
\description{
Set column description
}
\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotCols}}, \code{\link{annotRows<-}},
  \code{\link{annotRows}}, \code{\link{description}},
  \code{\link{main<-}}, \code{\link{main}},
  \code{\link{imputeData}}, \code{\link{imputeData<-}},
  \code{\link{qualityData}}, \code{\link{qualityData<-}},
  \code{\link{matrixData-class}}, \code{\link{matrixData}},
  \code{\link{names,matrixData-method}}
}
