% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\docType{methods}
\name{names,matrixData-method}
\alias{names,matrixData-method}
\title{Get names}
\usage{
\S4method{names}{matrixData}(x)
}
\arguments{
\item{x}{matrixData}
}
\description{
Get the column names of main and annotation columns.
}
\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotCols}}, \code{\link{annotRows<-}},
  \code{\link{annotRows}}, \code{\link{description<-}},
  \code{\link{description}}, \code{\link{main<-}},
  \code{\link{imputeData}}, \code{\link{imputeData<-}},
  \code{\link{qualityData}}, \code{\link{qualityData<-}},
  \code{\link{main}}, \code{\link{matrixData-class}},
  \code{\link{matrixData}}
}
