% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixData.R
\name{matrixData}
\alias{matrixData}
\title{matrixData constructor}
\usage{
matrixData(...)
}
\arguments{
\item{...}{\code{main}, \code{annotCols}, \code{annotRows}, \code{description}}
}
\description{
matrixData constructor
}
\seealso{
Other matrixData basic functions: \code{\link{annotCols<-}},
  \code{\link{annotCols}}, \code{\link{annotRows<-}},
  \code{\link{annotRows}}, \code{\link{description<-}},
  \code{\link{description}}, \code{\link{main<-}},
  \code{\link{main}}, \code{\link{matrixData-class}},
  \code{\link{names,matrixData-method}}
}
