% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.perseus}
\alias{read.perseus}
\title{Read Perseus matrix files}
\usage{
read.perseus(inFile)
}
\arguments{
\item{inFile}{path to input file}
}
\value{
DataFrame with additional 'annotationRows' attribute
}
\description{
Read the custom Perseus matrix file format *.txt into R.
}
\note{
Limitations to column names in R still apply. Column names valid
in Perseus, such as 'Column 1' will be changed to 'Column.1'
}
\examples{
testFile <- system.file('extdata', 'matrix1.txt', package='PerseusR')
df <- read.perseus(testFile)
}
\seealso{
\code{\link{write.perseus}}
}
