% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.perseus.default}
\alias{read.perseus.default}
\alias{read.perseus}
\alias{read.perseus.as.list}
\alias{read.perseus.as.matrixData}
\alias{read.perseus.as.ExpressionSet}
\title{Read Perseus matrix files}
\usage{
read.perseus.default(con, check = TRUE)

read.perseus.as.list(con, check = TRUE)

read.perseus.as.matrixData(con, check = TRUE)

read.perseus.as.ExpressionSet(con, check = TRUE)
}
\arguments{
\item{con}{A \code{\link{connection}} object or the path to input file}

\item{check}{Logical indicating whether to check for the validity of the exported object (slightly slower)}
}
\value{
DataFrame with additional 'annotationRows' attribute
}
\description{
Read the custom Perseus matrix file format *.txt into R.
}
\section{Functions}{
\itemize{
\item \code{read.perseus.as.list}: Difference between the mean and the median

\item \code{read.perseus.as.matrixData}: Difference between the mean and the median

\item \code{read.perseus.as.ExpressionSet}: Difference between the mean and the median
}}

\note{
Limitations to column names in R still apply. Column names valid
in Perseus, such as 'Column 1' will be changed to 'Column.1'

If the provided connection \code{con} is a character string, it will assumed
to be a file path. A \code{\link{connection}} which is not seekable (see \code{\link{isSeekable}})
will be written to a temporary file. Any connection will be closed when \code{read.perseus} exits.
\code{read.perseus.as.list}, \code{read.perseus.as.matrixData} and \code{read.perseus.as.ExpressionSet} are also available depending on the class desired as an output
}
\examples{

\dontrun{
mdata <- read.perseus(con=testFile)
}

}
\seealso{
\code{\link{write.perseus}}
}
