% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{singleChoiceParamInd}
\alias{singleChoiceParamInd}
\title{Single choice index}
\usage{
singleChoiceParamInd(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The selected index
}
\description{
Extract the index chosen in an \code{BoolParam}
}
\examples{
tmp <- tempfile(fileext = ".xml")
write('<SingleChoiceParam Name="test_single">\\n<Value>1</Value>\\n
<Values>\\n<Item>A</Item>\\n<Item>B</Item>\\n</Values>\\n</SingleChoiceParam>', file=tmp)
parameters <- parseParameters(tmp)
singleChoiceParamInd(parameters, "test_single")
}
