% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{singleChoiceParamValue}
\alias{singleChoiceParamValue}
\title{Single choice value}
\usage{
singleChoiceParamValue(parameters, name)
}
\arguments{
\item{parameters}{The parameters object (see \code{\link{parseParameters}})}

\item{name}{The name of the parameter}
}
\value{
The string representing the value
}
\description{
Extract the value selected in a \code{SingleChoiceParam}.
}
