% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.perseus}
\alias{write.perseus}
\title{Write matrixData in Perseus matrix format}
\usage{
write.perseus(mdata, con)
}
\arguments{
\item{mdata}{The matrixData}

\item{con}{A \code{\link{connection}} object or the path to output file}
}
\description{
Write a matrixData to file in the custom Perseus matrix file format.
}
\seealso{
\code{\link{read.perseus}} \code{\link{matrixData}}
}
